requirements.txt
flask
flask_cors
ultralytics
numpy
Pillow
opencv-python